+++ modules/cms/controllers/PostsController.php
@@ -7,6 +7,7 @@ use app\modules\cms\models\Post;
 use app\modules\cms\models\search\PostSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
 
 /**
@@ -32,6 +33,8 @@ class PostsController extends Controller
      */
     public function actionIndex()
     {
+		// Add access control
+		
         return $this->render('index', [
 			'posts' => Post::findRecentPosts(5),
 		]);
@@ -44,6 +47,8 @@ class PostsController extends Controller
      */
     public function actionView($id)
     {
+		// Add access control
+		
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -56,6 +61,10 @@ class PostsController extends Controller
      */
     public function actionDetails($id)
     {
+		if (!\Yii::$app->user->can('cmsPostsDetails')) {
+			throw new ForbiddenHttpException('You do not have privileges to view this content.');
+		}
+		
         return $this->render('details', [
             'model' => $this->findModel($id),
         ]);
@@ -68,6 +77,10 @@ class PostsController extends Controller
      */
     public function actionCreate()
     {
+		if (!\Yii::$app->user->can('cmsPostsCreate')) {
+			throw new ForbiddenHttpException('You do not have privileges to view this content.');
+		}
+		
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -87,6 +100,10 @@ class PostsController extends Controller
      */
     public function actionUpdate($id)
     {
+		if (!\Yii::$app->user->can('cmsPostsUpdate')) {
+			throw new ForbiddenHttpException('You do not have privileges to view this content.');
+		}
+		
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -106,6 +123,10 @@ class PostsController extends Controller
      */
     public function actionDelete($id)
     {
+		if (!\Yii::$app->user->can('cmsPostsDelete')) {
+			throw new ForbiddenHttpException('You do not have privileges to view this content.');
+		}
+		
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -117,6 +138,10 @@ class PostsController extends Controller
      */
     public function actionList()
     {
+		if (!\Yii::$app->user->can('cmsPostsList')) {
+			throw new ForbiddenHttpException('You do not have privileges to view this content.');
+		}
+		
         $searchModel = new PostSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
